+++ changelog.md
@@ -3,4 +3,5 @@
 ## [0.0.1] - 2025-08-04
 ### Added
 - AppBar with title
-- Card with TextFiled and Checkbox
+- Card with TextFiled and IconButton
+- Class Todo